requirements.txt
alembic==1.6.5
Babel==2.9.1
blinker==1.4
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-Principal==0.4.0
Flask-Security-Too==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
pkg-resources==0.0.0
psycopg2==2.9.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
WTForms==2.3.3
